reworked saves a bit
instead of saving entire LI array, it only saves the priority queue (needed to add character field to LI objects to make this possible, it should be helpful in the longrun anyways). also was able to move the loading process to be strictly in the SaveManager (instead of a combo effort between the SaveManager and the GameManager). i'm definitely missing other fixes but my brain is working overtime just to write these sentences